refactor: Repository 구조를 도메인별로 패키지화하여 정리
