+++ wecam-backend/src/main/java/org/example/wecambackend/repos/invitation/InvitationHistoryRepository.java
@@ -1,4 +1,4 @@
-package org.example.wecambackend.repos;
+package org.example.wecambackend.repos.invitation;
 
 import org.example.model.invitation.InvitationHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
