+++ wecam-backend/src/main/java/org/example/wecambackend/repos/school/SchoolRepository.java
@@ -1,4 +1,4 @@
-package org.example.wecambackend.repos;
+package org.example.wecambackend.repos.school;
 
 
 import org.example.model.University;
