+++ wecam-backend/src/main/java/org/example/wecambackend/repos/todo/TodoManagerRepository.java
@@ -1,4 +1,4 @@
-package org.example.wecambackend.repos;
+package org.example.wecambackend.repos.todo;
 
 import org.example.model.todo.Todo;
 import org.example.model.todo.TodoManager;
