+++ wecam-backend/src/main/java/org/example/wecambackend/repos/user/UserCustomRepository.java
@@ -1,4 +1,4 @@
-package org.example.wecambackend.repos;
+package org.example.wecambackend.repos.user;
 
 import org.example.wecambackend.dto.responseDTO.StudentSearchResponse;
