+++ wecam-backend/src/main/java/org/example/wecambackend/service/admin/AdminOrganizationService.java
@@ -14,10 +14,12 @@ import org.example.model.user.UserSignupInformation;
 import org.example.wecambackend.common.exceptions.BaseException;
 import org.example.wecambackend.common.response.BaseResponseStatus;
 import org.example.wecambackend.dto.responseDTO.*;
-import org.example.wecambackend.repos.*;
 import org.example.wecambackend.common.context.CouncilContextHolder;
+import org.example.wecambackend.repos.council.CouncilMemberRepository;
+import org.example.wecambackend.repos.council.CouncilRepository;
 import org.example.wecambackend.repos.organization.OrganizationRequestFileRepository;
 import org.example.wecambackend.repos.organization.OrganizationRequestRepository;
+import org.example.wecambackend.repos.user.UserSignupInformationRepository;
 import org.example.wecambackend.util.organization.OrganizationHierarchyUtil;
 import org.example.wecambackend.util.user.UserInfoExtractor;
 import org.springframework.beans.factory.annotation.Value;
