+++ wecam-backend/src/main/java/org/example/wecambackend/service/admin/AffiliationCertificationAdminService.java
@@ -17,9 +17,9 @@ import org.example.model.user.User;
 import org.example.model.affiliation.AffiliationCertification;
 import org.example.model.affiliation.AffiliationCertificationId;
 import org.example.model.enums.AuthenticationType;
-import org.example.wecambackend.repos.CouncilRepository;
+import org.example.wecambackend.repos.council.CouncilRepository;
 import org.example.wecambackend.repos.organization.OrganizationRepository;
-import org.example.wecambackend.repos.SchoolRepository;
+import org.example.wecambackend.repos.school.SchoolRepository;
 import org.example.wecambackend.repos.affiliation.AffiliationCertificationRepository;
 import org.example.wecambackend.repos.affiliation.AffiliationFileRepository;
 import org.example.wecambackend.service.admin.common.AdminFileStorageService;
