+++ wecam-backend/src/main/java/org/example/wecambackend/service/admin/CouncilDepartmentService.java
@@ -7,8 +7,8 @@ import org.example.model.council.CouncilDepartment;
 import org.example.model.council.CouncilDepartmentRole;
 import org.example.wecambackend.common.exceptions.BaseException;
 import org.example.wecambackend.common.response.BaseResponseStatus;
-import org.example.wecambackend.repos.CouncilDepartmentRepository;
-import org.example.wecambackend.repos.CouncilDepartmentRoleRepository;
+import org.example.wecambackend.repos.council.CouncilDepartmentRepository;
+import org.example.wecambackend.repos.council.CouncilDepartmentRoleRepository;
 import org.example.wecambackend.service.admin.common.EntityFinderService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
