+++ wecam-backend/src/main/java/org/example/wecambackend/service/admin/StudentService.java
@@ -9,9 +9,9 @@ import org.example.wecambackend.dto.projection.OrganizationNameLevelDto;
 import org.example.wecambackend.dto.responseDTO.UserSummaryResponse;
 import org.example.wecambackend.dto.responseDTO.StudentSearchResponse;
 import org.example.wecambackend.dto.requestDTO.StudentSearchRequest;
-import org.example.wecambackend.repos.CouncilRepository;
-import org.example.wecambackend.repos.UserInformationRepository;
-import org.example.wecambackend.repos.UserRepository;
+import org.example.wecambackend.repos.council.CouncilRepository;
+import org.example.wecambackend.repos.user.UserInformationRepository;
+import org.example.wecambackend.repos.user.UserRepository;
 import org.example.wecambackend.service.admin.common.EntityFinderService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
