+++ wecam-backend/src/main/java/org/example/wecambackend/service/auth/AuthService.java
@@ -19,10 +19,10 @@ import org.example.wecambackend.dto.requestDTO.StudentRegisterRequest;
 import org.example.model.user.User;
 import org.example.model.user.UserPrivate;
 import org.example.model.user.UserSignupInformation;
-import org.example.wecambackend.repos.CouncilMemberRepository;
-import org.example.wecambackend.repos.UserPrivateRepository;
-import org.example.wecambackend.repos.UserRepository;
-import org.example.wecambackend.repos.UserSignupInformationRepository;
+import org.example.wecambackend.repos.council.CouncilMemberRepository;
+import org.example.wecambackend.repos.user.UserPrivateRepository;
+import org.example.wecambackend.repos.user.UserRepository;
+import org.example.wecambackend.repos.user.UserSignupInformationRepository;
 import org.example.wecambackend.util.PhoneEncryptor;
 import org.flywaydb.core.internal.util.StringUtils;
 import org.springframework.data.redis.core.RedisTemplate;
