+++ wecam-backend/src/main/java/org/example/wecambackend/service/client/Affiliation/AffiliationService.java
@@ -16,9 +16,11 @@ import org.example.model.affiliation.AffiliationCertificationId;
 import org.example.model.enums.AuthenticationStatus;
 import org.example.model.enums.AuthenticationType;
 import org.example.model.enums.OcrResult;
-import org.example.wecambackend.repos.*;
 import org.example.wecambackend.repos.affiliation.AffiliationCertificationRepository;
 import org.example.wecambackend.repos.organization.OrganizationRepository;
+import org.example.wecambackend.repos.school.SchoolRepository;
+import org.example.wecambackend.repos.user.UserRepository;
+import org.example.wecambackend.repos.user.UserSignupInformationRepository;
 import org.example.wecambackend.service.client.common.filesave.FilePath;
 import org.example.wecambackend.service.client.common.filesave.FileStorageService;
 import org.springframework.stereotype.Service;
