+++ wecam-backend/src/main/java/org/example/wecambackend/service/client/MyPageService.java
@@ -5,7 +5,7 @@ import org.example.model.user.UserSignupInformation;
 import org.example.wecambackend.common.exceptions.BaseException;
 import org.example.wecambackend.common.response.BaseResponseStatus;
 import org.example.wecambackend.dto.requestDTO.MyPageOrganizationEditRequest;
-import org.example.wecambackend.repos.UserSignupInformationRepository;
+import org.example.wecambackend.repos.user.UserSignupInformationRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -15,9 +15,9 @@ import org.example.wecambackend.dto.responseDTO.MyPageResponse;
 import org.example.model.organization.Organization;
 import org.example.model.user.User;
 import org.example.model.user.UserInformation;
-import org.example.wecambackend.repos.UserInformationRepository;
-import org.example.wecambackend.repos.UserPrivateRepository;
-import org.example.wecambackend.repos.UserRepository;
+import org.example.wecambackend.repos.user.UserInformationRepository;
+import org.example.wecambackend.repos.user.UserPrivateRepository;
+import org.example.wecambackend.repos.user.UserRepository;
 import org.springframework.stereotype.Service;
 import org.example.wecambackend.util.PhoneEncryptor;
